refactor(settings): 拆出文件管理器设置页签状态 hook 与容器
- 新增 settings/fileManager/：useFileManagerSettings（状态/持久化/下载目录预览）、FileManagerTabPane（页签容器）
- SettingsModal.tsx 1204 -> 718 行
